+++ controlsdk_streaming/src/main/java/org/btelman/controlsdk/streaming/video/processors/BaseVideoProcessor.kt
@@ -1,4 +1,4 @@
-package org.btelman.controlsdk.streaming.components
+package org.btelman.controlsdk.streaming.video.processors
 
 /**
  * Handles core logic of processing the video and usually sending it out to a service
